requirements.txt
bs4
flask
pymongo
psycopg2-binary
pandas
numpy
pyyaml
redis
wheel
requests
lxml
